Move statement to ast
