+++ librlox/src/ast/mod.rs
@@ -1,4 +1,5 @@
 pub mod expression;
+pub mod statement;
 
 #[cfg(test)]
 mod tests;
