+++ librlox/src/ast/tests/ast.rs
@@ -1,4 +1,5 @@
 use crate::ast::expression::{Expr, MultiplicationExpr, PrimaryExpr, UnaryExpr};
+use crate::ast::statement::Stmt;
 use crate::scanner::tokens::{Token, TokenType, Value};
 use std::convert::TryFrom;
 use std::option::Option;
@@ -26,4 +27,17 @@ fn test_expression_formatter_should_pretty_print_an_ast() {
         "(* (- 123) (Grouping 45.7))".to_string(),
         format!("{}", expr)
     )
+}
+
+#[test]
+fn test_statement_formatter_should_pretty_print_an_ast() {
+    let expr = Stmt::Expression(Expr::Unary(UnaryExpr::Minus(Box::new(Expr::Primary(
+        PrimaryExpr::try_from(Token::new(
+            TokenType::Literal,
+            Option::Some(Value::Number(123.0)),
+        ))
+        .unwrap(),
+    )))));
+
+    assert_eq!("(Expression (- 123))".to_string(), format!("{}", expr))
 }
