+++ librlox/src/interpreter/tests/statement/mod.rs
@@ -1,7 +1,7 @@
 use crate::ast::expression::{Expr, PrimaryExpr};
+use crate::ast::statement::Stmt;
 use crate::interpreter::InterpreterMut;
 use crate::interpreter::StatefulInterpreter;
-use crate::parser::statement::Stmt;
 
 #[test]
 fn expression_stmt_should_return_ok() {
