+++ librlox/src/parser/mod.rs
@@ -1,6 +1,5 @@
 mod combinators;
 pub mod expression_parser;
-pub mod statement;
 pub mod statement_parser;
 
 #[cfg(test)]
