+++ librlox/src/parser/tests/statement_parser.rs
@@ -1,25 +1,11 @@
 extern crate parcel;
-use crate::ast::expression::{Expr, PrimaryExpr, UnaryExpr};
-use crate::parser::statement::Stmt;
+use crate::ast::expression::{Expr, PrimaryExpr};
+use crate::ast::statement::Stmt;
 use crate::parser::statement_parser::statements;
 use crate::scanner::tokens::{Token, TokenType, Value};
 use parcel::prelude::v1::*;
 use parcel::MatchStatus;
-use std::convert::TryFrom;
 use std::option::Option;
-
-#[test]
-fn test_expression_formatter_should_pretty_print_an_ast() {
-    let expr = Stmt::Expression(Expr::Unary(UnaryExpr::Minus(Box::new(Expr::Primary(
-        PrimaryExpr::try_from(Token::new(
-            TokenType::Literal,
-            Option::Some(Value::Number(123.0)),
-        ))
-        .unwrap(),
-    )))));
-
-    assert_eq!("(Expression (- 123))".to_string(), format!("{}", expr))
-}
 
 #[test]
 fn test_parser_can_parse_declaration_stmt() {
